ksmbd: don't check negotiate context size in smb2_check_message
NegotiateContextOffset and NegotiateContextCount do not always guarantee
zero in a dialect that is not smb 3.11. In the smb2.1 connection from
some app, kernel oops occurs by setting the two values to a value other
than zero and sending them. Because it is difficult to determine
the dialect before smb2 negotiate in smb2_check_message.

Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>